+++ src/checkers/wp_plugins/mod.rs
@@ -2,4 +2,5 @@
 //! They have nothing special, they are standard checkers, but they are
 //! similar and grouping them in a module helps keeping a clean architecture.
 
+pub mod revslider;
 pub mod yoast_seo;